game class
