+++ utility.js
@@ -1,0 +1,8 @@
+export let COLORS = ['blue','green','red'];
+export const BUTTON_HEIGHT =40;
+export const BUTTON_WIDTH =80;
+export const CANVAS_HEIGHT = 800;
+export const CANVAS_WIDTH = 800;
+export let speed = 8;
+export let points =0;
+export const vel = 1;
